+++ npm/webpack-preprocessor/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@cypress/webpack-preprocessor-v5.15.3](https://github.com/cypress-io/cypress/compare/@cypress/webpack-preprocessor-v5.15.2...@cypress/webpack-preprocessor-v5.15.3) (2022-11-02)
+
+
+### Bug Fixes
+
+* Ensure patched merge-source-map is included in package ([#24490](https://github.com/cypress-io/cypress/issues/24490)) ([493d90c](https://github.com/cypress-io/cypress/commit/493d90c33082565a93418f39388daab348c175cf))
+
 # [@cypress/webpack-preprocessor-v5.15.2](https://github.com/cypress-io/cypress/compare/@cypress/webpack-preprocessor-v5.15.1...@cypress/webpack-preprocessor-v5.15.2) (2022-11-01)